Add setup for pytest
This commit adds a `tests` directory with the necessary boilerplate to
import our main module into the tests.

The `context.py` file modifies Python's load path to include the parent
dir, which contains our main module. This allows us to import that
module using a relative import in the test file.