Configure linter
This commit adds a config file for flake8, in order to configure it to
skip linting the `context.py` in the test directory, which does some
importing funny business that the linter disagrees with. (If we decide
later that it is right and we should change how we're handling that, we
can remove that exclusion from the config.)

It also makes the GHA step fail instead of warn, to ensure we fix
our errors for now.