Only run `main` when executing as a script
This commit adds a conditional check for calling that main function,
that appears to be standard in Python for ensuring that `main` is not
called when importing the file into another file (like we do in tests).
With this conditional, `main` will only be executed if the file is
invoked as a script with `python artifact_calc.py`.

Without this, when running the tests with `pytest --cov=artifact_calc`,
the coverage report was showing as 100%, even if no functions from the
module were called in the tests, because `main` was getting executed on
import, and the main function currently exercises the whole file.